requirements.txt
# Python version: 3.12
# Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.7.4
pydantic-settings==2.1.0
jinja2>=3.1.0

# LangChain & LangGraph
langgraph==0.1.9
langchain==0.2.5
langchain-openai==0.1.10
langchain-core==0.2.19
langsmith>=0.1.0

# LLM & AI
openai==1.26.0

# Email handling
imap-tools==1.6.0
email-validator==2.1.0
aiosmtplib>=3.0.0

# HTTP & async
httpx==0.25.2
aiohttp==3.9.1

# Data & database
sqlalchemy==2.0.23
alembic==1.13.1

# Environment & configuration
python-dotenv==1.0.0

# Scheduling
schedule==1.2.1
APScheduler==3.10.4

# Logging
loguru==0.7.2

# Vector search and embeddings
faiss-cpu>=1.8.0
numpy>=1.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
ruff==0.1.8
mypy==1.7.1

# Type hints
types-python-dateutil==2.8.19
types-pytz==2023.3.0.0
